+++ plugins/polio/js/src/domains/LQAS-IM/IM/utils.ts
@@ -8,14 +8,13 @@ import {
     LqasImCampaign,
     LqasImDistrictData,
 } from '../../../constants/types';
+import { IN_SCOPE } from '../shared/constants';
 import {
-    IM_PASS,
-    IM_FAIL,
-    IM_WARNING,
-    imNfmKeys,
-    IN_SCOPE,
-} from '../shared/constants';
-import { findDataForShape, findScopeIds, makeLegendItem } from '../../../utils';
+    LegendItem,
+    findDataForShape,
+    findScopeIds,
+    makeLegendItem,
+} from '../../../utils';
 import { OK_COLOR, WARNING_COLOR, FAIL_COLOR } from '../../../styles/constants';
 import {
     sortGraphKeys,
@@ -27,6 +26,7 @@ import {
 
 import { determineStatusForDistrict as lqasDistrictStatus } from '../LQAS/utils';
 import { LQASIMType } from '../shared/types/types';
+import { IM_FAIL, IM_PASS, IM_WARNING, imNfmKeys } from './constants';
 
 export const determineStatusForDistrict = district => {
     if (!district) return null;
@@ -61,7 +61,7 @@ export const makeImMapLegendItems =
         imData: Record<string, ConvertedLqasImData>,
         campaign: string | undefined,
         round: number,
-    ): { label: string; value: string; color: unknown }[] => {
+    ): LegendItem[] => {
         const [passed, failed, disqualified] = getImStatsForRound(
             imData,
             campaign,
@@ -69,17 +69,17 @@ export const makeImMapLegendItems =
         );
         const passedLegendItem = makeLegendItem({
             color: OK_COLOR,
-            value: passed?.length,
+            value: passed ? `${passed?.length}` : '',
             label: formatMessage(MESSAGES['1imOK']),
         });
         const disqualifiedLegendItem = makeLegendItem({
             color: WARNING_COLOR,
-            value: disqualified?.length,
+            value: disqualified ? `${disqualified?.length}` : '',
             label: formatMessage(MESSAGES['2imWarning']),
         });
         const failedLegendItem = makeLegendItem({
             color: FAIL_COLOR,
-            value: failed?.length,
+            value: failed ? `${failed?.length}` : '',
             label: formatMessage(MESSAGES['3imFail']),
         });
