+++ plugins/polio/js/src/domains/LQAS-IM/LQAS/LqasAfroOverview/ListView/LqasAfroList.tsx
@@ -32,7 +32,7 @@ import {
 } from '../hooks/useAfroMapShapes';
 import { getRound } from '../utils';
 import { LqasAfroOverviewContext } from '../Context/LqasAfroOverviewContext';
-import { HasLocationIcon } from './HasLocationIcon';
+import { HasLocationIcon } from '../../../shared/HasLocationIcon';
 import { IN_SCOPE } from '../../../shared/constants';
 import { LQAS_FAIL, LQAS_PASS } from '../../constants';
