+++ plugins/polio/js/src/domains/LQAS-IM/shared/DatesIgnored.tsx
@@ -22,24 +22,16 @@ export const DatesIgnored: FunctionComponent<Props> = ({ data, campaign }) => {
         ? data?.day_country_not_found[currentCountryName] ?? {}
         : {};
     const datesArray = Object.keys(datesIgnored);
-    return (
+    return campaign && datesArray.length !== 0 ? (
         <>
-            {campaign && datesArray.length !== 0 && (
-                <>
-                    <Typography variant="h6">
-                        {`${formatMessage(MESSAGES.datesIgnored)}:`}
-                    </Typography>
-                    {datesArray.map(d => (
-                        <Box mr={1} mb={1} display="inline-block" key={d}>
-                            <Chip
-                                label={d}
-                                variant="outlined"
-                                color="secondary"
-                            />
-                        </Box>
-                    ))}
-                </>
-            )}
+            <Typography variant="h6">
+                {`${formatMessage(MESSAGES.datesIgnored)}:`}
+            </Typography>
+            {datesArray.map(d => (
+                <Box mr={1} mb={1} display="inline-block" key={d}>
+                    <Chip label={d} variant="outlined" color="secondary" />
+                </Box>
+            ))}
         </>
-    );
+    ) : null;
 };
